Provide event to Dialog `onClose` callback.
This will give Dialog consumers additional context
when deciding whether to prevent closing the dialog.

An example for where this is useful is when an action is
triggered from a dialog that results in a transient notification
(e.g., a snackbar) that includes dismiss and/or action buttons.
Clicking on a button inside the notification should not close the
dialog, even if the notification itself is not a child of the
dialog in the DOM.